requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
django-filter
django-crispy-forms
django-admin-interface==0.26.0
django-colorfield==0.11.0
django-jet-reboot==1.3.5
crispy-bootstrap5
django-widget-tweaks
drf-yasg
celery
redis
gunicorn
django-environ
reportlab
openpyxl
django-import-export
python-barcode
Pillow
requests